+++ include/CutPlotter.hpp
@@ -18,7 +18,7 @@
 
 using namespace std;
 
-typedef function<double ()> eventContainerFunction;
+typedef function<float ()> eventContainerFunction;
 
 class CutPlotter : BasePlotter {
