+++ include/EventSelecter.hpp
@@ -10,7 +10,7 @@
 
 using namespace std;
 
-typedef function<double ()> eventContainerFunction;
+typedef function<float ()> eventContainerFunction;
 typedef function<bool ()> comparisonFunction;
 
 // enum class ComparisonType { GREATER_THAN, SMALLER_THAN, ABS_GREATER_THAN, ABS_SMALLER_THAN, EQUAL }; // defined in VariableDictionary
