+++ src/Met.cpp
@@ -3,7 +3,7 @@
 
 Met::Met() { }
 
-Met::Met(double pt, double phi) 
+Met::Met(float pt, float phi) 
     : Particle( pt, 0, phi, 0) { }
 
 Met::Met(const TLorentzVector& p4) 
@@ -11,7 +11,7 @@ Met::Met(const TLorentzVector& p4)
 
 Met::~Met() { }
 
-void Met::set(double pt, double phi) 
+void Met::set(float pt, float phi) 
 {
     Particle::set( pt, 0, phi, 0);
 }
