+++ src/Particle.cpp
@@ -7,7 +7,7 @@
 
 Particle::Particle() { }
 
-Particle::Particle(double pt, double eta, double phi, double mass)
+Particle::Particle(float pt, float eta, float phi, float mass)
 {
     p4.SetPtEtaPhiM(pt, eta, phi, mass);
 }
@@ -16,7 +16,7 @@ Particle::Particle(const TLorentzVector& p4Vec) : p4( p4Vec ) { }
 
 Particle::~Particle() { }
     
-void Particle::set(double pt, double eta, double phi, double mass) 
+void Particle::set(float pt, float eta, float phi, float mass) 
 {
     p4.SetPtEtaPhiM(pt, eta, phi, mass);
 }  
@@ -26,48 +26,48 @@ const TLorentzVector& Particle::getLorentzVector() const
     return p4;
 }
 
-double Particle::pt() const
+float Particle::pt() const
 {
     return p4.Pt();
 }
-double Particle::eta() const
+float Particle::eta() const
 {
     return p4.Eta();
 }
-double Particle::phi() const
+float Particle::phi() const
 {
     return p4.Phi();
 }
-double Particle::mass() const
+float Particle::mass() const
 {
     return p4.M();
 }
     
-double Particle::dR(const Particle& particle) const 
+float Particle::dR(const Particle& particle) const 
 {
     const TLorentzVector& lorentzV = particle.getLorentzVector();
     return p4.DeltaR( lorentzV );
 }
 
-double Particle::dPhi(const Particle& particle) const 
+float Particle::dPhi(const Particle& particle) const 
 {
     const TLorentzVector& lorentzV = particle.getLorentzVector();
     return p4.DeltaPhi( lorentzV );
 }
 
-double Particle::dEta(const Particle& particle) const 
+float Particle::dEta(const Particle& particle) const 
 {
     const TLorentzVector& lorentzV = particle.getLorentzVector();
     return p4.Eta() - lorentzV.Eta();
 }
 
-double Particle::mpp(const Particle& particle) const 
+float Particle::mpp(const Particle& particle) const 
 {
     const TLorentzVector& lorentzV = particle.getLorentzVector();
     return (p4+lorentzV).M() ;
 }
 
-double Particle::ptpp(const Particle& particle) const 
+float Particle::ptpp(const Particle& particle) const 
 {
     const TLorentzVector& lorentzV = particle.getLorentzVector();
     return (p4+lorentzV).Pt() ;
